chore(utxo-lib): move psbt determinisitic funcs to testutil
BG-76314
